refactor(react,website): remove Svg from SideNavigation components
replacing usage of Svg for icons from the iconography

BREAKING CHANGE: svgName and svgClass props no longer exist, use icon prop instead
