+++ packages/website/src/examples/SideNavigation/loading.example.tsx
@@ -1,0 +1,12 @@
+import {SideNavigation, SideNavigationLoadingItem, SideNavigationMenuSection} from '@coveord/plasma-react';
+import {CoveoIconSize16Px} from '@coveord/plasma-react-icons';
+
+export default () => (
+    <SideNavigation>
+        <SideNavigationMenuSection title="Loading Section" icon={CoveoIconSize16Px}>
+            <SideNavigationLoadingItem className="mod-width-30" />
+            <SideNavigationLoadingItem className="mod-width-50" />
+            <SideNavigationLoadingItem className="mod-width-40" />
+        </SideNavigationMenuSection>
+    </SideNavigation>
+);
