+++ packages/website/src/pages/navigation/SideNavigation.tsx
@@ -1,80 +1,7 @@
 import {PageLayout} from '../../building-blocs/PageLayout';
-
-const loading = `
-    import {
-        SideNavigation,
-        SideNavigationLoadingItem,
-        SideNavigationMenuSection,
-    } from '@coveord/plasma-react';
-    
-    export default () => (
-        <SideNavigation>
-            <SideNavigationMenuSection title="Loading Section" svgName={'coveoIcon16px'}>
-                <SideNavigationLoadingItem className="mod-width-30" />
-                <SideNavigationLoadingItem className="mod-width-50" />
-                <SideNavigationLoadingItem className="mod-width-40" />
-            </SideNavigationMenuSection>
-        </SideNavigation>
-    );
-`;
-const collapsible = `
-    import {
-        SideNavigation,
-        SideNavigationItem,
-        SideNavigationMenuSection,
-    } from '@coveord/plasma-react';
-
-    export default () => {
-        const [isExpanded, setIsExpanded] = React.useState(true);
-        return (
-            <SideNavigation>
-                <SideNavigationMenuSection
-                    title="Collapsible Section"
-                    svgName={'coveoIcon16px'}
-                    expandable
-                    expanded={isExpanded}
-                    onClick={() => setIsExpanded(!isExpanded)}
-                >
-                    <SideNavigationItem>
-                        <a href="http://coveo.com" title="Link to Coveo">
-                            Link to Coveo
-                        </a>
-                    </SideNavigationItem>
-                    <SideNavigationItem>
-                        <a href="http://coveo.com" title="Link to Coveo">
-                            Another link to Coveo
-                        </a>
-                    </SideNavigationItem>
-                </SideNavigationMenuSection>
-            </SideNavigation>
-        );
-    };
-`;
-
-const code = `
-    import {
-        SideNavigation,
-        SideNavigationItem,
-        SideNavigationMenuSection,
-    } from '@coveord/plasma-react';
-
-    export default () => (
-        <SideNavigation>
-        <SideNavigationMenuSection title="Regular Section" svgName={'coveoIcon16px'}>
-            <SideNavigationItem isActive>
-                <a href="http://coveo.com" title="Link to Coveo">
-                    Link to Coveo
-                </a>
-            </SideNavigationItem>
-            <SideNavigationItem>
-                <a href="http://coveo.com" title="Link to Coveo">
-                    Another link to Coveo
-                </a>
-            </SideNavigationItem>
-        </SideNavigationMenuSection>
-    </SideNavigation>
-    );
-`;
+import collapsible from '../../examples/SideNavigation/collapsible.example.tsx';
+import loading from '../../examples/SideNavigation/loading.example.tsx';
+import code from '../../examples/SideNavigation/main.example.tsx';
 
 export const SideNavigationExamples = () => (
     <PageLayout
